refactor: extract shared helpers into a package-level utils/
core/utils.py was 641 lines across four unrelated concerns and core/fits.py is
generic I/O — neither is a command implementation. Mirrors pfb_imaging/utils/
and is what the broken `from spimple.utils import ...` lines were reaching for.

Also drops the only direct python-casacore import: set_wcs now converts the MS
epoch with astropy.time, pinned by four regression vectors captured from the
casacore implementation. Fixes ORIGIN, which claimed "pfb-imaging".
