+++ src/spimple/utils/beam.py
@@ -5,144 +5,11 @@ from africanus.rime.dask import beam_cube_dde as beam_cube_dde_dask
 from africanus.rime.fast_beam_cubes import beam_cube_dde
 from africanus.util.numba import jit
 from astropy.io import fits
-from astropy.wcs import WCS
 import dask.array as da
 from daskms import xds_from_ms, xds_from_table
-from ducc0.fft import c2r, good_size, r2c
 import numpy as np
-import ray
-from reproject import reproject_interp
-from reproject.mosaicking import find_optimal_celestial_wcs
-from scipy import ndimage
-from scipy.interpolate import RegularGridInterpolator
-import xarray as xr
 
-from spimple.core.fits import data_from_header, load_fits
-
-iFs = np.fft.ifftshift
-Fs = np.fft.fftshift
-
-
-def Gaussian2D(xin, yin, GaussPar=(1.0, 1.0, 0.0), normalise=True, nsigma=5):
-    """
-    xin         - grid of x coordinates
-    yin         - grid of y coordinates
-    GaussPar    - (emaj, emin, pa) with emaj/emin in units x and pa in radians.
-    normalise   - normalise kernel to have volume 1
-    nsigma      - compute kernel out to this many sigmas
-
-    Note - the rotation matrix is defined as
-
-    [[np.sin(PA), -np.cos(PA)],
-     [np.cos(PA), np.sin(PA)]]
-
-    instead of
-
-    [[np.cos(PA), -np.sin(PA)],
-     [np.sin(PA), np.cos(PA)]]
-
-    with t = pi/2 - pa
-
-    for compatibility with fits
-    """
-    Smaj, Smin, PA = GaussPar
-    A = np.array([[1.0 / Smaj**2, 0], [0, 1.0 / Smin**2]])
-    R = np.array([[np.sin(PA), -np.cos(PA)], [np.cos(PA), np.sin(PA)]])
-    A = np.dot(np.dot(R.T, A), R)
-    sOut = xin.shape
-    # only compute the result out to 5 * emaj
-    extent = (nsigma * Smaj) ** 2
-    xflat = xin.squeeze()
-    yflat = yin.squeeze()
-    idx, idy = np.where(xflat**2 + yflat**2 <= extent)
-    x = np.array([xflat[idx, idy].ravel(), yflat[idx, idy].ravel()])
-    R = np.einsum("nb,bc,cn->n", x.T, A, x)
-    # need to adjust for the fact that GaussPar corresponds to FWHM
-    fwhm_conv = 2 * np.sqrt(2 * np.log(2))
-    tmp = np.exp(-fwhm_conv * R)
-    gausskern = np.zeros(xflat.shape, dtype=np.float64)
-    gausskern[idx, idy] = tmp
-
-    if normalise:
-        gausskern /= np.sum(gausskern)
-    return np.ascontiguousarray(gausskern.reshape(sOut), dtype=np.float64)
-
-
-def get_padding_info(nx, ny, pfrac):
-    npad_x = int(pfrac * nx)
-    nfft = good_size(nx + npad_x, True)
-    npad_xl = (nfft - nx) // 2
-    npad_xr = nfft - nx - npad_xl
-
-    npad_y = int(pfrac * ny)
-    nfft = good_size(ny + npad_y, True)
-    npad_yl = (nfft - ny) // 2
-    npad_yr = nfft - ny - npad_yl
-    padding = ((0, 0), (npad_xl, npad_xr), (npad_yl, npad_yr))
-    unpad_x = slice(npad_xl, -npad_xr)
-    unpad_y = slice(npad_yl, -npad_yr)
-    return padding, unpad_x, unpad_y
-
-
-def convolve2gaussres(image, xx, yy, gaussparf, nthreads, gausspari=None, pfrac=0.5, norm_kernel=False):
-    """
-    Convolves the image to a specified resolution.
-
-    Parameters
-    ----------
-    Image - (nband, nx, ny) array to convolve
-    xx/yy - coordinates on the grid in the same units as gaussparf.
-    gaussparf - tuple containing Gaussian parameters of desired resolution
-                (emaj, emin, pa).
-    gausspari - initial resolution . By default it is assumed that the image
-                is a clean component image with no associated resolution.
-                If beampari is specified, it must be a tuple containing gausspars
-                for each imaging band in the same format.
-    nthreads - number of threads to use for the FFT's.
-    pfrac - padding used for the FFT based convolution. Will pad by pfrac/2 on
-            both sides of image
-    """
-    nband, nx, ny = image.shape
-    padding, unpad_x, unpad_y = get_padding_info(nx, ny, pfrac)
-    ax = (1, 2)  # axes over which to perform fft
-    lastsize = ny + np.sum(padding[-1])
-
-    gausskern = Gaussian2D(xx, yy, gaussparf, normalise=norm_kernel)
-    gausskern = np.pad(gausskern[None], padding, mode="constant")
-    gausskernhat = r2c(iFs(gausskern, axes=ax), axes=ax, forward=True, nthreads=nthreads, inorm=0)
-
-    image = np.pad(image, padding, mode="constant").astype(np.float64)
-    imhat = r2c(iFs(image, axes=ax), axes=ax, forward=True, nthreads=nthreads, inorm=0)
-
-    # convolve to desired resolution
-    if gausspari in [None, ()]:
-        imhat *= gausskernhat
-    else:
-        for i in range(nband):
-            thiskern = Gaussian2D(xx, yy, gausspari[i], normalise=norm_kernel).astype(np.float64)
-            thiskern = np.pad(thiskern[None], padding, mode="constant")
-            thiskernhat = r2c(
-                iFs(thiskern, axes=ax),
-                axes=ax,
-                forward=True,
-                nthreads=nthreads,
-                inorm=0,
-            )
-
-            if not np.all(np.isnan(thiskernhat)):
-                convkernhat = np.where(np.abs(thiskernhat) > 1e-10, gausskernhat / thiskernhat, 0.0)
-            else:
-                print("Nan values have been encountered. Subverting RuntimeWarning")
-                convkernhat = np.zeros_like(thiskernhat).astype("complex")
-
-            imhat[i] *= convkernhat[0]
-
-    image = Fs(
-        c2r(imhat, axes=ax, forward=False, lastsize=lastsize, inorm=2, nthreads=nthreads),
-        axes=ax,
-    )[:, unpad_x, unpad_y]
-
-    return image, gausskern[:, unpad_x, unpad_y]
+from spimple.utils.fits import load_fits
 
 
 @jit(nopython=True, nogil=True, cache=True)
@@ -446,196 +313,3 @@ def interpolate_beam(ll, mm, freqs, opts):
     # swap source and freq axes and reshape to image shape
     beam_source = np.transpose(beam_image, axes=(1, 0))
     return beam_source.squeeze().reshape((freqs.size, *ll.shape))
-
-
-def mosaic_info(im_list, oname, ref_image=None):
-    if ref_image is not None:
-        raise NotImplementedError("Reference image not supported yet")
-    wcss = []
-    out_names = []
-    freqs = []
-    flatfreqs = []
-    shapes = []
-    basename = oname.removesuffix(".fits")
-    for imnum, im in enumerate(im_list):
-        hdr = fits.getheader(im)
-        nu = data_from_header(hdr, axis=3)[0]
-        freqs.append(nu)
-        flatfreqs.extend(nu)
-        wcsi = WCS(hdr).dropaxis(-1).dropaxis(-1)
-        wcss.append(wcsi)
-        nchan = hdr["NAXIS3"]
-        ncorr = hdr["NAXIS4"]
-        shapes.append((hdr["NAXIS1"], hdr["NAXIS2"]))
-        out_names.extend([f"{basename}_im{imnum}_pol{c}_ch{f}.zarr" for c in range(ncorr) for f in range(nchan)])
-    nu = np.array(flatfreqs)
-    ufreqs = np.unique(nu)
-    # get domain of intrinsic image
-    ref_wcs, shape_out = find_optimal_celestial_wcs(list(zip(shapes, wcss, strict=False)), projection="SIN")
-    ref_wcs.array_shape = (shape_out[0], shape_out[1])
-
-    return ref_wcs, ufreqs, out_names
-
-
-@ray.remote
-def project(im, imnum, ref_wcs, beam, oname, method="interp"):
-    if method != "interp":
-        raise NotImplementedError("Only 'interp' method supported for now")
-
-    # output shape
-    nxo, nyo = ref_wcs.array_shape
-
-    # interpolate beam
-    bds = xr.open_zarr(beam, chunks=None)
-    beam = bds.BEAM.values
-    l_beam = bds.l_beam.values
-    m_beam = bds.m_beam.values
-    bfreq = bds.chan.values
-
-    # make the power beam
-    pbeam = ((beam[0] * beam[0].conj()).real + (beam[-1] * beam[-1].conj()).real) / 2.0
-
-    # this is cheap, evaluation is more expensive
-    beamo = RegularGridInterpolator(
-        (bfreq, l_beam, m_beam),
-        pbeam,
-        bounds_error=False,
-        fill_value=None,
-        method="linear",
-    )
-
-    cell_x, cell_y = ref_wcs.wcs.cdelt
-    l_im = (-(nxo) // 2 + np.arange(nxo)) * cell_x
-    m_im = (-(nyo) // 2 + np.arange(nyo)) * cell_y
-    im_coords = {
-        "l": ("l", l_im),
-        "m": ("m", m_im),
-    }
-
-    image = fits.getdata(im)
-    ncorr, nchan, nx, ny = image.shape
-    hdr = fits.getheader(im)
-    freq, _ = data_from_header(hdr, axis=3)
-    wcs = WCS(hdr).dropaxis(-1).dropaxis(-1)
-    cxi, cyi = wcs.wcs.cdelt
-    nx, ny = wcs.array_shape
-    l = (-(nx) // 2 + np.arange(nx)) * cxi
-    m = (-(ny) // 2 + np.arange(ny)) * cyi
-    ll, mm = np.meshgrid(l, m, indexing="ij")
-
-    basename = oname.removesuffix(".fits")
-    # TODO - make these in parallel
-    for c in range(ncorr):
-        for f in range(nchan):
-            bdata = np.zeros((nx, ny), dtype=np.float64)
-            beami = beamo((freq[c], ll, mm))
-            step = 25
-            angles = np.linspace(0, 359, step)
-            for angle in angles:
-                bdata += ndimage.rotate(beami, angle, reshape=False, order=1, mode="nearest")
-            bdata /= angles.size
-            pbeam, footprint = reproject_interp(
-                (bdata, wcs),
-                ref_wcs,
-                shape_out=(nxo, nyo),
-                block_size="auto",
-                parallel=4,
-            )
-            mask = footprint > 0
-            pbeam[~mask] = 0
-            pdata, _ = reproject_interp(
-                (image[c, f], wcs),
-                ref_wcs,
-                shape_out=(nxo, nyo),
-                block_size="auto",
-                parallel=4,
-            )
-
-            im_attrs = {
-                "freq": freq[c],
-            }
-
-            data_vars = {
-                "IMAGE": (("l", "m"), pdata),
-                "BEAM": (("l", "m"), pbeam),
-                "MASK": (("l", "m"), mask),
-            }
-
-            ds = xr.Dataset(data_vars, coords=im_coords, attrs=im_attrs).chunk({
-                "l": 512,
-                "m": 512,
-            })
-
-            ds_name = f"{basename}_im{imnum}_pol{c}_ch{f}.zarr"
-            ds.to_zarr(ds_name, compute=True, mode="w", consolidated=False)
-
-    return imnum
-
-
-@ray.remote
-def stitch_images(freq, im_list, eta=1e-3):
-    # get all datasets in current plane
-    xds = []
-    for im in im_list:
-        ds = xr.open_zarr(im, chunks=None, consolidated=False)
-        if ds.freq == freq:
-            xds.append(ds)
-
-    # accumulate
-    nx = xds[0].l.size
-    ny = xds[0].m.size
-    y = np.zeros((nx, ny))
-    for ds in xds:
-        mask = ds.MASK.values
-        beam = ds.BEAM.values
-        image = ds.IMAGE.values
-        y[mask] += beam[mask] * image[mask]
-    ds = ds.drop_vars("IMAGE")
-
-    def hess(x):
-        out = np.zeros((nx, ny))
-        for ds in xds:
-            mask = ds.MASK.values
-            beam = ds.BEAM.values
-            res = beam**2 * x
-            out[mask] += res[mask]
-        return out + eta * x
-
-    image, info = conjugate_gradient(hess, y, max_iter=10)
-
-    weight = np.zeros((nx, ny))
-    for ds in xds:
-        mask = ds.MASK.values
-        beam = ds.BEAM.values
-        weight[mask] += beam[mask] ** 2
-    weight += eta
-    return image, weight, info, freq
-
-
-def conjugate_gradient(A, b, x0=None, tol=1e-6, max_iter=100, report=20):
-    n = b.shape
-    x = np.zeros(n) if x0 is None else x0.copy()
-    r = A(x) - b
-    p = -r
-    rsold = np.vdot(r, r)
-    rs0 = rsold
-    if rs0 < tol:
-        return x, 0  # already at minimum
-    for i in range(max_iter):
-        Ap = A(p)
-        alpha = rsold / np.vdot(p, Ap)
-        x = x + alpha * p
-        r = r + alpha * Ap
-        rsnew = np.vdot(r, r)
-
-        if np.sqrt(rsnew) < tol:
-            break
-
-        beta = rsnew / rsold
-        p = beta * p - r
-        rsold = rsnew
-
-        if i % report == 0:
-            print(f"At {i} norm frac = {rsnew / rs0}")
-
-    return x, i
